feat: Update storage configuration and implement price handling in cents
- Enhanced S3 client configuration in StorageConfig to use ApacheHttpClient with timeouts.
- Modified BatchRequest, BatchResponse, ProductBatchRequest, and StockSummaryResponse to use Long for cost and selling prices, representing values in cents.
- Updated Batch entity to store cost and selling prices as Long.
- Refactored BatchService to generate unique batch codes and handle price conversions.
- Added migration script to convert existing price fields from DECIMAL to BIGINT (cents).
- Updated integration and unit tests to reflect changes in price handling and batch code generation.
- Adjusted application configuration for storage settings and added test configurations.

+++ src/main/java/br/com/stockshift/service/ReportService.java
@@ -44,7 +44,7 @@ public class ReportService {
 
         BigDecimal totalStockValue = allBatches.stream()
                 .filter(b -> b.getCostPrice() != null)
-                .map(b -> b.getCostPrice().multiply(BigDecimal.valueOf(b.getQuantity())))
+                .map(b -> BigDecimal.valueOf(b.getCostPrice()).multiply(BigDecimal.valueOf(b.getQuantity())))
                 .reduce(BigDecimal.ZERO, BigDecimal::add);
 
         long pendingMovements = stockMovementRepository.findByTenantIdAndStatus(tenantId, MovementStatus.PENDING).size();
@@ -120,7 +120,7 @@ public class ReportService {
 
         BigDecimal totalValue = batches.stream()
                 .filter(b -> b.getCostPrice() != null)
-                .map(b -> b.getCostPrice().multiply(BigDecimal.valueOf(b.getQuantity())))
+                .map(b -> BigDecimal.valueOf(b.getCostPrice()).multiply(BigDecimal.valueOf(b.getQuantity())))
                 .reduce(BigDecimal.ZERO, BigDecimal::add);
 
         LocalDate nearestExpiration = batches.stream()
@@ -143,7 +143,7 @@ public class ReportService {
 
     private StockReportResponse batchToReport(Batch batch) {
         BigDecimal totalValue = batch.getCostPrice() != null ?
-                batch.getCostPrice().multiply(BigDecimal.valueOf(batch.getQuantity())) :
+                BigDecimal.valueOf(batch.getCostPrice()).multiply(BigDecimal.valueOf(batch.getQuantity())) :
                 BigDecimal.ZERO;
 
         return StockReportResponse.builder()
